+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Fixed: regression from 1.2.5 on calc() on multiple lines
+  (@MoOx)
+
 # 1.2.5 - 2016-08-20
 
 - Fixed: security issue due to the usage of ``eval()``.